Module_6_Multi_Agents/A2A ADK MCP/requirements.txt
fastapi>=0.95.0
google-generativeai==0.8.4
google-adk==0.1.0
google-auth==2.38.0
google-cloud-dlp==3.28.0
google-cloud-secret-manager==2.23.2
pandas==2.2.3
langchain==0.3.19
langchain-google-genai==2.0.10
langchain-community==0.3.18
mcp==1.6.0
loguru
deprecated